feat(build): Build each aur package in a separate stage.
The stages copy forward a `/next` file to prevent them from all running
in parallel and overloading the build box. If buildkit implements any
sort of concurrency limiting, then this hack could be removed. It might
also be possible to implement this hack with more than one chain of
builds, but that is just too messy to deal with.